Validate in createOrUpdate and delete function

+++ spring-boot-web/src/main/java/com/javaweb/api/admin/BuildingAPI.java
@@ -17,6 +17,8 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,6 +33,7 @@ public class BuildingAPI {
     @PostMapping
     private ResponseEntity<?> createOrUpdateBuilding(@Valid @RequestBody BuildingDTO buildingDTO,
                                                   BindingResult bindingResult) {
+        ResponseDTO responseDTO = new ResponseDTO();
         try{
             if (bindingResult.hasErrors()) {
                 List<String> errorMessages = bindingResult.getFieldErrors()
@@ -38,17 +41,18 @@ public class BuildingAPI {
                         .map(FieldError::getDefaultMessage)
                         .collect(Collectors.toList());
 
-                ResponseDTO responseDTO = new ResponseDTO();
                 responseDTO.setMessage("Validation failed");
                 responseDTO.setDetail(errorMessages);
                 return ResponseEntity.badRequest().body(responseDTO);
             }
             // neu dung thi //xuong service -> xuong repo -> save vao db
-            ResponseDTO responseDTO = buildingService.createOrUpdate(buildingDTO);
+            responseDTO = buildingService.createOrUpdate(buildingDTO);
             return ResponseEntity.status(HttpStatus.CREATED).body(responseDTO);
         }
         catch (Exception e){
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+            responseDTO.setMessage("Internal server error");
+            responseDTO.setDetail(Collections.singletonList(e.getMessage()));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseDTO);
         }
     }
 
@@ -59,14 +63,46 @@ public class BuildingAPI {
     }
 
     @PutMapping("/assignment")
-    private ResponseDTO updateAssignmentBuilding(@RequestBody AssignmentBuildingDTO assignmentBuildingDTO) {
-        ResponseDTO responseDTO = buildingService.updateAssignmentModal(assignmentBuildingDTO);
-        return responseDTO;
+    private ResponseEntity<?> updateAssignmentBuilding(@Valid @RequestBody AssignmentBuildingDTO assignmentBuildingDTO,
+                                                       BindingResult bindingResult) {
+        ResponseDTO responseDTO = new ResponseDTO();
+        try {
+            if (bindingResult.hasErrors()) {
+                List<String> errorMessages = bindingResult.getFieldErrors()
+                        .stream()
+                        .map(FieldError::getDefaultMessage)
+                        .collect(Collectors.toList());
+
+                responseDTO.setMessage("Validation failed");
+                responseDTO.setDetail(errorMessages);
+                return ResponseEntity.badRequest().body(responseDTO);
+            }
+            // neu dung thi //xuong service -> xuong repo -> save vao db
+            responseDTO = buildingService.updateAssignmentModal(assignmentBuildingDTO);
+            return ResponseEntity.status(HttpStatus.CREATED).body(responseDTO);
+        } catch (Exception e) {
+            responseDTO.setMessage("Internal server error");
+            responseDTO.setDetail(Collections.singletonList(e.getMessage()));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseDTO);
+        }
     }
 
     @DeleteMapping("/{ids}")
-    private ResponseDTO deleteBuilding(@PathVariable List<Long> ids) {
-        ResponseDTO responseDTO = buildingService.deleteBuildings(ids);
-        return responseDTO;
+    private ResponseEntity<?> deleteBuilding(@PathVariable List<Long> ids) {
+        ResponseDTO responseDTO = new ResponseDTO();
+        try {
+            if (ids.toArray().length == 0) {
+                responseDTO.setMessage("Validation failed");
+                responseDTO.setDetail(Collections.singletonList("Please select at least one building to delete"));
+                return ResponseEntity.badRequest().body(responseDTO);
+            }
+
+            responseDTO = buildingService.deleteBuildings(ids);
+            return ResponseEntity.ok().body(responseDTO);
+        } catch (Exception e) {
+            responseDTO.setMessage("Internal server error");
+            responseDTO.setDetail(Collections.singletonList(e.getMessage()));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseDTO);
+        }
     }
 }
